% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosphere_list_params.R
\name{geosphere_list_params}
\alias{geosphere_list_params}
\title{List Parameters for a Given Resource}
\usage{
geosphere_list_params(resource_id)
}
\arguments{
\item{resource_id}{The ID of the resource to list parameters for.}
}
\value{
A vector of parameter names.
}
\description{
This function lists the parameters available for a given resource.
}
