% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosphere_download_from_filelisting.R
\name{geosphere_download_from_schema}
\alias{geosphere_download_from_schema}
\title{Download File Using a Predefined Geosphere Schema}
\usage{
geosphere_download_from_schema(
  dest_dir,
  resource_id,
  params,
  base_data_url = "https://public.hub.geosphere.at/datahub/resources/",
  user_agent = NULL,
  overwrite = FALSE,
  verbose = TRUE,
  timeout_seconds = 300
)
}
\arguments{
\item{dest_dir}{Directory to save the file.}

\item{resource_id}{The key in the internal schemas list corresponding to the dataset
(e.g., "spartacus-v2-1d-1km"). This is also the \code{resource_id}
used in the URL.}

\item{params}{A named list of parameter values to fill into the templates
(e.g., \code{list(year = 2020, variable_type = "TX")}).}

\item{base_data_url}{Base URL for Geosphere data.}

\item{user_agent}{Custom User-Agent. If \code{NULL}, a default package user agent is used.}

\item{overwrite}{Logical, whether to overwrite existing files.}

\item{verbose}{Logical, for verbose output.}

\item{timeout_seconds}{Request timeout.}
}
\value{
Path to the downloaded file or \code{NA_character_} on failure.
}
\description{
Constructs the filename and subpath based on a predefined internal schema for a given
Geosphere Austria \code{resource_id} and a set of parameters, then downloads the file.
This is an internal helper function for downloading known datasets.
}
\keyword{internal}
